fingerprints/pyproject.toml
[project]
name = "fingerprints"
version = "0.1.0"
description = "Lightweight SHA-256 fingerprint harvester for open-source LLM releases on HuggingFace"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "huggingface-hub>=0.23",
    "click>=8.1",
    "rich>=13.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "python-json-logger>=2.0",
    "requests>=2.31",
    "pyarrow>=14.0",
]

[project.scripts]
fingerprints = "fingerprints.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fingerprints"]
